Fix widget color on the welcome popup to be black instead of white
Also, PR#108 deleted the WidgetManager.js ... I didn't catch it before, but when I tried to test this CSS I had issues because this was missing, so readding it.

The reason why it wasn't an issue on dev or production is because they already had the file in a temporary store and used it even though the jar no longer had it.